Refactor new RPC into several files (#1007)
* From/TryFrom starknet api types

* Add deploy account

* Modify gitignore

* Deploy account and invoke function

* Change into_iter to iter

* Update .gitignore


* change to try_from

* Move functions to its respective files

* WIP Added SIR support to SNRPC

* Implemented state reader for sir

* WIP Transaction

* WIP SiR execution

* Fixed rpc sir execute_tx

* Fix clippy

* Import last version of sn_api

* Formatting

* Test

* Test try from

* Delete test

* Fix format

* Fix test

* local test cases

* specify block manually

* add test_case for blockifier

* use more recent txs

* dont keep cached responses

* add failing tx on blockifier

* more tests

* Fix clippy

* fix bug

* tests

* Replaced try_from with from_invoke_transaction

* infer version

* Fix version

* Changed test_try_from_invoke

* Ignore test_recent_tx

* ignore failing tests

* Refactor tx deser, (un)ignore tests

* sorted assert and fee threshold

* refactor rpc state reader

* fixes

---------
