+++ rpc_state_reader_sn_api/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 ureq = { version = "2.7.1", features = ["json"] }
-serde = { version = "=1.0.171", features = ["derive"] }
+serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0", features = [
     "arbitrary_precision",
     "raw_value",
