+++ rpc_state_reader_sn_api/tests/blockifier_tests.rs
@@ -1,0 +1,318 @@
+use blockifier::{
+    block_context::BlockContext,
+    execution::{contract_class::ContractClass, entry_point::CallInfo},
+    state::{
+        cached_state::{CachedState, GlobalContractCache},
+        state_api::{StateReader, StateResult},
+    },
+    transaction::{
+        account_transaction::AccountTransaction, objects::TransactionExecutionInfo,
+        transactions::ExecutableTransaction,
+    },
+};
+use blockifier::{
+    execution::contract_class::{
+        ContractClass as BlockifierContractClass, ContractClassV0, ContractClassV0Inner,
+    },
+    transaction::transactions::InvokeTransaction,
+};
+use cairo_lang_starknet::{
+    casm_contract_class::CasmContractClass, contract_class::ContractClass as SierraContractClass,
+};
+use cairo_vm::types::program::Program;
+use pretty_assertions_sorted::{assert_eq, assert_eq_sorted};
+use rpc_state_reader_sn_api::rpc_state::*;
+use rpc_state_reader_sn_api::utils;
+use starknet::core::types::ContractClass as SNContractClass;
+use starknet_api::{
+    block::BlockNumber,
+    contract_address,
+    core::{ClassHash, CompiledClassHash, ContractAddress, Nonce, PatriciaKey},
+    hash::{StarkFelt, StarkHash},
+    patricia_key, stark_felt,
+    state::StorageKey,
+    transaction::{Transaction as SNTransaction, TransactionHash},
+};
+use test_case::test_case;
+
+use std::{collections::HashMap, sync::Arc};
+
+pub struct RpcStateReader(RpcState);
+
+impl StateReader for RpcStateReader {
+    fn get_storage_at(
+        &mut self,
+        contract_address: ContractAddress,
+        key: StorageKey,
+    ) -> StateResult<StarkFelt> {
+        Ok(self.0.get_storage_at(&contract_address, &key))
+    }
+
+    fn get_nonce_at(&mut self, contract_address: ContractAddress) -> StateResult<Nonce> {
+        Ok(Nonce(self.0.get_nonce_at(&contract_address)))
+    }
+
+    fn get_class_hash_at(&mut self, contract_address: ContractAddress) -> StateResult<ClassHash> {
+        Ok(self.0.get_class_hash_at(&contract_address))
+    }
+
+    /// Returns the contract class of the given class hash.
+    fn get_compiled_contract_class(
+        &mut self,
+        class_hash: &ClassHash,
+    ) -> StateResult<ContractClass> {
+        Ok(match self.0.get_contract_class(class_hash) {
+            SNContractClass::Legacy(compressed_legacy_cc) => {
+                let as_str = utils::decode_reader(compressed_legacy_cc.program).unwrap();
+                let program = Program::from_bytes(as_str.as_bytes(), None).unwrap();
+                let entry_points_by_type = utils::map_entry_points_by_type_legacy(
+                    compressed_legacy_cc.entry_points_by_type,
+                );
+                let inner = Arc::new(ContractClassV0Inner {
+                    program,
+                    entry_points_by_type,
+                });
+                BlockifierContractClass::V0(ContractClassV0(inner))
+            }
+            SNContractClass::Sierra(flattened_sierra_cc) => {
+                let middle_sierra: utils::MiddleSierraContractClass = {
+                    let v = serde_json::to_value(flattened_sierra_cc).unwrap();
+                    serde_json::from_value(v).unwrap()
+                };
+                let sierra_cc = SierraContractClass {
+                    sierra_program: middle_sierra.sierra_program,
+                    contract_class_version: middle_sierra.contract_class_version,
+                    entry_points_by_type: middle_sierra.entry_points_by_type,
+                    sierra_program_debug_info: None,
+                    abi: None,
+                };
+                let casm_cc = CasmContractClass::from_contract_class(sierra_cc, false).unwrap();
+                BlockifierContractClass::V1(casm_cc.try_into().unwrap())
+            }
+        })
+    }
+
+    /// Returns the compiled class hash of the given class hash.
+    fn get_compiled_class_hash(&mut self, class_hash: ClassHash) -> StateResult<CompiledClassHash> {
+        Ok(CompiledClassHash(
+            self.0
+                .get_class_hash_at(&ContractAddress(class_hash.0.try_into().unwrap()))
+                .0,
+        ))
+    }
+}
+
+pub fn execute_tx(
+    tx_hash: &str,
+    network: RpcChain,
+    block_number: BlockNumber,
+) -> (
+    TransactionExecutionInfo,
+    TransactionTrace,
+    RpcTransactionReceipt,
+) {
+    let tx_hash = tx_hash.strip_prefix("0x").unwrap();
+
+    // Instantiate the RPC StateReader and the CachedState
+    let rpc_reader = RpcStateReader(RpcState::new_infura(network, block_number.into()));
+    let gas_price = rpc_reader.0.get_gas_price(block_number.0).unwrap();
+
+    // Get values for block context before giving ownership of the reader
+    let chain_id = rpc_reader.0.get_chain_name();
+    let RpcBlockInfo {
+        block_number,
+        block_timestamp,
+        sequencer_address,
+        ..
+    } = rpc_reader.0.get_block_info();
+
+    // Get transaction before giving ownership of the reader
+    let tx_hash = TransactionHash(stark_felt!(tx_hash));
+    let sn_api_tx = rpc_reader.0.get_transaction(&tx_hash);
+
+    let trace = rpc_reader.0.get_transaction_trace(&tx_hash);
+    let receipt = rpc_reader.0.get_transaction_receipt(&tx_hash);
+
+    // Create state from RPC reader
+    let global_cache = GlobalContractCache::default();
+    let mut state = CachedState::new(rpc_reader, global_cache);
+
+    let fee_token_address =
+        contract_address!("049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7");
+
+    const N_STEPS_FEE_WEIGHT: f64 = 0.01;
+    let vm_resource_fee_cost = Arc::new(HashMap::from([
+        ("n_steps".to_string(), N_STEPS_FEE_WEIGHT),
+        ("output_builtin".to_string(), 0.0),
+        ("pedersen_builtin".to_string(), N_STEPS_FEE_WEIGHT * 32.0),
+        ("range_check_builtin".to_string(), N_STEPS_FEE_WEIGHT * 16.0),
+        ("ecdsa_builtin".to_string(), N_STEPS_FEE_WEIGHT * 2048.0),
+        ("bitwise_builtin".to_string(), N_STEPS_FEE_WEIGHT * 64.0),
+        ("ec_op_builtin".to_string(), N_STEPS_FEE_WEIGHT * 1024.0),
+        ("poseidon_builtin".to_string(), N_STEPS_FEE_WEIGHT * 32.0),
+        (
+            "segment_arena_builtin".to_string(),
+            N_STEPS_FEE_WEIGHT * 10.0,
+        ),
+        ("keccak_builtin".to_string(), N_STEPS_FEE_WEIGHT * 2048.0), // 2**11
+    ]));
+
+    let block_context = BlockContext {
+        chain_id,
+        block_number,
+        block_timestamp,
+        sequencer_address,
+        fee_token_address,
+        vm_resource_fee_cost,
+        gas_price,
+        invoke_tx_max_n_steps: 1_000_000,
+        validate_max_n_steps: 1_000_000,
+        max_recursion_depth: 500,
+    };
+
+    // Map starknet_api transaction to blockifier's
+    let blockifier_tx = match sn_api_tx {
+        SNTransaction::Invoke(tx) => {
+            let invoke = InvokeTransaction { tx, tx_hash };
+            AccountTransaction::Invoke(invoke)
+        }
+        _ => unimplemented!(),
+    };
+
+    (
+        blockifier_tx
+            .execute(&mut state, &block_context, true, true)
+            .unwrap(),
+        trace,
+        receipt,
+    )
+}
+
+#[test]
+fn test_get_gas_price() {
+    let block = BlockValue::Number(BlockNumber(169928));
+    let rpc_state = RpcState::new_infura(RpcChain::MainNet, block);
+
+    let price = rpc_state.get_gas_price(169928).unwrap();
+    assert_eq!(price, 22804578690);
+}
+
+#[test]
+fn blockifier_test_recent_tx() {
+    let (tx_info, trace, receipt) = execute_tx(
+        "0x05d200ef175ba15d676a68b36f7a7b72c17c17604eda4c1efc2ed5e4973e2c91",
+        RpcChain::MainNet,
+        BlockNumber(169928),
+    );
+
+    let TransactionExecutionInfo {
+        execute_call_info,
+        actual_fee,
+        ..
+    } = tx_info;
+
+    let CallInfo {
+        vm_resources,
+        inner_calls,
+        ..
+    } = execute_call_info.unwrap();
+
+    assert_eq!(actual_fee.0, receipt.actual_fee);
+    assert_eq!(
+        vm_resources,
+        trace
+            .function_invocation
+            .as_ref()
+            .unwrap()
+            .execution_resources
+    );
+    assert_eq!(
+        inner_calls.len(),
+        trace
+            .function_invocation
+            .as_ref()
+            .unwrap()
+            .internal_calls
+            .len()
+    );
+}
+
+#[test_case(
+    "0x05d200ef175ba15d676a68b36f7a7b72c17c17604eda4c1efc2ed5e4973e2c91",
+    169928, // real block 169929
+    RpcChain::MainNet
+)]
+#[test_case(
+    "0x0528ec457cf8757f3eefdf3f0728ed09feeecc50fd97b1e4c5da94e27e9aa1d6",
+    169928, // real block 169929
+    RpcChain::MainNet
+)]
+#[test_case(
+    "0x0737677385a30ec4cbf9f6d23e74479926975b74db3d55dc5e46f4f8efee41cf",
+    169928, // real block 169929
+    RpcChain::MainNet
+    => ignore["resource mismatch"]
+)]
+#[test_case(
+    "0x026c17728b9cd08a061b1f17f08034eb70df58c1a96421e73ee6738ad258a94c",
+    169928, // real block 169929
+    RpcChain::MainNet
+)]
+#[test_case(
+    // review later
+    "0x0743092843086fa6d7f4a296a226ee23766b8acf16728aef7195ce5414dc4d84",
+    186548, // real block     186549
+    RpcChain::MainNet
+    => ignore["resource mismatch"]
+)]
+#[test_case(
+    // fails in blockifier too
+    "0x00724fc4a84f489ed032ebccebfc9541eb8dc64b0e76b933ed6fc30cd6000bd1",
+    186551, // real block     186552
+    RpcChain::MainNet
+)]
+fn blockifier_test_case_tx(hash: &str, block_number: u64, chain: RpcChain) {
+    let (tx_info, trace, receipt) = execute_tx(hash, chain, BlockNumber(block_number));
+
+    let TransactionExecutionInfo {
+        execute_call_info,
+        actual_fee,
+        ..
+    } = tx_info;
+
+    let CallInfo {
+        vm_resources,
+        inner_calls,
+        ..
+    } = execute_call_info.unwrap();
+
+    let actual_fee = actual_fee.0;
+    if receipt.actual_fee != actual_fee {
+        let diff = 100 * receipt.actual_fee.abs_diff(actual_fee) / receipt.actual_fee;
+
+        if diff >= 5 {
+            assert_eq!(
+                actual_fee, receipt.actual_fee,
+                "actual_fee mismatch differs from the baseline by more than 5% ({diff}%)",
+            );
+        }
+    }
+
+    assert_eq_sorted!(
+        vm_resources,
+        trace
+            .function_invocation
+            .as_ref()
+            .unwrap()
+            .execution_resources
+    );
+    assert_eq!(
+        inner_calls.len(),
+        trace
+            .function_invocation
+            .as_ref()
+            .unwrap()
+            .internal_calls
+            .len()
+    );
+}
